config-generator-beta/requirements.txt
# For testing
pyfakefs
PyYAML

# For the generator scripts
absl-py
Jinja2
requests